chore: move Encrypter to its own file
