+++ src/utils/helpers/index.js
@@ -1,0 +1,5 @@
+const Encrypter = require('./encrypter');
+
+module.exports = {
+  Encrypter
+};
